applied patch ID: 3606722

+++ src/ofmt.cpp
@@ -675,7 +675,7 @@ OFmtI( ostream* os, SizeT offs, SizeT r, int w, int d, char f,
       (*os) << oct << setw(w) << setfill(f) << (*this)[ i];
   else if ( oMode == BIN)
     for( SizeT i=offs; i<endEl; ++i)
-      (*os) << binstr((*this)[ i], w);
+      (*os) << binstr((int)(*this)[ i], w);
 //       (*os) << binstr((std::bitset<32>)i, w);
   else if ( oMode == HEX)
     for( SizeT i=offs; i<endEl; ++i)
@@ -713,8 +713,8 @@ OFmtI( ostream* os, SizeT offs, SizeT r, int w, int d, char f,
 //       if (w > 32) 
 // 	(*os) << binstr((std::bitset<32>)i >> 32, w - 32);
 //       (*os) << binstr((std::bitset<32>)i, w <= 32 ? w : 32);
-      if (w > 32) (*os) << binstr((*this)[ i] >> 32, w - 32);
-      (*os) << binstr((*this)[ i], w <= 32 ? w : 32);
+      if (w > 32) (*os) << binstr((int)(*this)[ i] >> 32, w - 32);
+      (*os) << binstr((int)(*this)[ i], w <= 32 ? w : 32);
     }
   else if ( oMode == HEX)
     for( SizeT i=offs; i<endEl; ++i)
